refactor domain product graph route
